grt: refactor code to not use duplicated static_casts

+++ src/grt/src/Grid.cpp
@@ -173,32 +173,26 @@ interval<int>::type Grid::computeTileReduceInterval(
 
 int Grid::computeTileReduce(const odb::Rect& obs,
                             const odb::Rect& tile,
-                            int track_space,
+                            double track_space,
                             bool first,
                             odb::dbTechLayerDir direction)
 {
   int reduce = -1;
   if (direction == odb::dbTechLayerDir::VERTICAL) {
     if (obs.xMin() >= tile.xMin() && obs.xMax() <= tile.xMax()) {
-      reduce = ceil(static_cast<double>(std::abs(obs.xMax() - obs.xMin()))
-                    / track_space);
+      reduce = ceil(std::abs(obs.xMax() - obs.xMin()) / track_space);
     } else if (first) {
-      reduce = ceil(static_cast<double>(std::abs(tile.xMax() - obs.xMin()))
-                    / track_space);
+      reduce = ceil(std::abs(tile.xMax() - obs.xMin()) / track_space);
     } else {
-      reduce = ceil(static_cast<double>(std::abs(obs.xMax() - tile.xMin()))
-                    / track_space);
+      reduce = ceil(std::abs(obs.xMax() - tile.xMin()) / track_space);
     }
   } else {
     if (obs.yMin() >= tile.yMin() && obs.yMax() <= tile.yMax()) {
-      reduce = ceil(static_cast<double>(std::abs(obs.yMax() - obs.yMin()))
-                    / track_space);
+      reduce = ceil(std::abs(obs.yMax() - obs.yMin()) / track_space);
     } else if (first) {
-      reduce = ceil(static_cast<double>(std::abs(tile.yMax() - obs.yMin()))
-                    / track_space);
+      reduce = ceil(std::abs(tile.yMax() - obs.yMin()) / track_space);
     } else {
-      reduce = ceil(static_cast<double>(std::abs(obs.yMax() - tile.yMin()))
-                    / track_space);
+      reduce = ceil(std::abs(obs.yMax() - tile.yMin()) / track_space);
     }
   }
